feat: 개인 지표별 안정화 값 저장하도록 함. 각 지표별 안정점수를 책정해서 안정하다고 판단 시 지표별 유효성을 검사하고 후보값으로 등록, 그렇게 모인 후보값들 중에서 밀집구간의 중앙값을 안정화 값으로 사용.

+++ app/src/main/java/com/leejang/sleeptandard/Screen/ExperimentScreen.kt
@@ -381,7 +381,13 @@ private fun RespirationGraphCard(data: PpgRespirationGraphData) {
                 .fillMaxWidth()
                 .height(220.dp)
         )
-        StatusLine("분석 구간", "%.1f / %.1f sec".format(data.windowSeconds, data.minimumWindowSeconds))
+        StatusLine(
+            "분석 구간",
+            "%.1f / %.1f sec".format(
+                data.windowSeconds,
+                data.minimumWindowSeconds.toDouble()
+            )
+        )
         StatusLine("호흡수", data.calculatedRrBpm?.let { "%.1f bpm".format(it) } ?: "--")
         StatusLine("품질", formatNullable(data.qualityScore, 3))
     }
